move generate_topology to correct location
